Move distance formatting out of utils
