+++ event-details-boundary/TravelEstimates.tsx
@@ -20,7 +20,7 @@ import { BodyText, Caption, CaptionTitle, Headline } from "@components/Text"
 import { openEventLocationInMaps } from "./Event"
 import { Ionicon, RoundedIonicon } from "@components/common/Icons"
 import { AppStyles } from "@lib/AppColorStyle"
-import { compactFormatDistance } from "@lib/utils/DistanceFormatting"
+import { compactFormatDistance } from "@lib/DistanceFormatting"
 import { dayjs } from "TiFShared/lib/Dayjs"
 import duration from "dayjs/plugin/duration"
 import { AvatarMapMarkerView } from "@components/AvatarMapMarker"
