+++ lib/DistanceFormatting.ts
